MdeModulePkg/Variable/VariableRuntimeDxe: Add Bool PcdDriverConvertsFvbFuncPointers
VariableRuntimeDxe is unconditionally converting the pointers to the Fvb protocol
functions even if the platform drivers are already converting the pointers
themselves.  This leads to a double pointer conversion and can cause unexpected
runtime behaviour depending on the memory layout of the platform.

Since we don't want to break legacy and out of tree platforms we add a Bool flag
that defaults to the existing behaviour but allows platforms to only use the
pointer conversion being done in their driver implementation that produces the
FVB Protocol.

Signed-off-by: Jon Nettleton <jon@solid-run.com>